part 4 done

+++ part4/exercises_4.1_to_4.23/controller/blogs.js
@@ -1,8 +1,9 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
 
 blogsRouter.get('/', async (request, response) => {
-  const allBlogs = await Blog.find({})
+  const allBlogs = await Blog.find({}).populate('user', {username: 1, name: 1, id: 1})
   response.json(allBlogs)
 })
   
@@ -13,15 +14,38 @@ blogsRouter.post('/', async ( request, response) => {
   if(!request.body.hasOwnProperty('title') || !request.body.hasOwnProperty('url')){
     response.status(400).send({ error: "No title or URL"})
   } else {
-  const blog = new Blog(request.body)
+
+  if (!request.user) {    
+    response.status(401).json({ error: 'token invalid' })  
+  }
+
+  const user = await User.findById(request.user)
+  const blog = new Blog({
+    ...request.body,
+    user: user.id
+  })
+  user.blogs = user.blogs.concat(blog.id)
+  await user.save()
+
   response.status(201).json(await blog.save())
   }
 })
 
 blogsRouter.delete('/:id', async ( request, response ) => {
   try{
-    await Blog.findByIdAndDelete(request.params.id)
+    if (!request.user) {    
+      response.status(401).json({ error: 'token invalid' })  
+    }
+    const blog = await Blog.findById(request.params.id)
+    console.log("Blog:", blog)
+    console.log("Blog user ID", blog.user.toString())
+    console.log("Token ID", request.user)
+    if(blog.user.toString() === request.user){
+      await Blog.findByIdAndDelete(request.params.id)
     response.status(204).end()
+    } else {
+      response.status(401).json({ error: 'user not authorised' })
+    }
   } catch(exception){
     console.log(exception)
   }
